Sort morphemes when serializing them
To ensure better consistency and to make the resulting file prettier,
sort the morphemes of each word by span before outputting them.

Note that this may make the result of some (arguably improperly coded)
operations inconsistent when running on a lexicon created in memory vs.
running on the same lexicon after saving and reloading, because the
morphemes may get reordered in between. But it ensures the results will
be always identical after any fresh reload, no matter how the file was
created.