Test morpheme identity with `is` in simple seg
We want to insert a ` + ` into the simple segmentation whenever there
is a break between two morphemes. The break can be detected using
either equality or object identity, since when two morphemes are
identical, they necessarilly have identical spans and therefore the
breaks between them are not well-defined anyway, but using object
identity is theoretically faster.