Add a method for obtaining morpheme forms
I'm not very happy with the differences between the API for morphemes
and for lexemes – lexemes are identified by their ID and their features
are obtained by calling metods of `SegLex` with the ID as an argument.
Morphemes are objects (namedtuples) and their properties are accessed
directly. But it is what we have now.