Parse the prefix/root/suffix info from data
The segmentation information in MorphoLEX contains morpheme types, but
we ignored them so far. Parse the information in a more structured way,
so that we get this info out of it.

While doing so, I discovered tens of structural errors in the
annotation, which prevent successful parse. I fix them in my local copy
of the data. Maybe I'll prepare a "patch" in the future?