Distinguish more annot layers vs homonymy on load
When loading files, a single lexeme may have multiple records
corresponding to different annotation layers, and this case must be
distinguished from homonyms. Sometimes it can't be done without user's
manual intervention (e.g. add an extra key to the lexical features
distinguishing the homonyms), but usually it is possible.