Detect changes in overlap when writing simple seg
The simple segmentation in the 4th column of the file should contain
a ` + ` whenever there is a change in the morpheme. But when morphemes
overlap, it is hard to say what counts as a change. Up until now, that
was non-deterministic, causing issues with round-trips of the files.
Hopefully, detecting any change in morpheme lists at a given position
makes the process fully deterministic.